+++ CHANGELOG.md
@@ -2,6 +2,9 @@ Changelog
 =========
 
 ## 0.3.0 (2013-05-xx)
+* [BC break] `GenericOAuth2ResourceOwner::getAccessToken` now returns an array
+  instead of a string. This array contains the access token and its 'expires_in'
+  value, along with any other parameters returned from the authentication provider
 * [BC break] Removed `AdvancedPathUserResponse` & `AdvancedUserResponseInterface`
 * [BC break] Added `UserResponseInterface#getEmail()` & `UserResponseInterface#getProfilePicture()`
   methods